requirements.txt
telethon
discord.py
pyyaml
